Remove old parser and make the new parser work
