+++ letsencrypt-nginx/letsencrypt_nginx/parser.py
@@ -55,12 +55,12 @@ class NginxParser(object):
                 if _is_include_directive(entry):
                     # Parse the top-level included file
                     self._parse_recursively(entry[1])
-                elif entry[0] == ['http'] or entry[0] == ['server']:
+                elif len(entry) > 0 and (entry[0] == ['http'] or entry[0] == ['server']):
                     # Look for includes in the top-level 'http'/'server' context
                     for subentry in entry[1]:
                         if _is_include_directive(subentry):
                             self._parse_recursively(subentry[1])
-                        elif entry[0] == ['http'] and subentry[0] == ['server']:
+                        elif entry[0] == ['http'] and len(subentry) > 0 and subentry[0] == ['server']:
                             # Look for includes in a 'server' context within
                             # an 'http' context
                             for server_entry in subentry[1]:
@@ -102,7 +102,7 @@ class NginxParser(object):
             srv = servers[filename]  # workaround undefined loop var in lambdas
 
             # Find all the server blocks
-            _do_for_subarray(tree, lambda x: x[0] == ['server'],
+            _do_for_subarray(tree, lambda x: _is_directive(x, ['server']),
                              lambda x: srv.append(x[1]))
 
             # Find 'include' statements in server blocks and append their trees
@@ -289,7 +289,7 @@ class NginxParser(object):
 
         """
         _do_for_subarray(self.parsed[filename],
-                         lambda x: x[0] == ['http'],
+                         lambda x: _is_directive(x, ['http']),
                          lambda x: x[1].insert(0, directives))
 
     def get_all_certs_keys(self):
@@ -308,9 +308,9 @@ class NginxParser(object):
             tup = [None, None, vhost.filep]
             if vhost.ssl:
                 for directive in vhost.raw:
-                    if directive[0] == 'ssl_certificate':
+                    if _is_directive(directive, 'ssl_certificate'):
                         tup[0] = directive[1]
-                    elif directive[0] == 'ssl_certificate_key':
+                    elif _is_directive(directive, 'ssl_certificate_key'):
                         tup[1] = directive[1]
             if tup[0] is not None and tup[1] is not None:
                 c_k.add(tuple(tup))
@@ -463,15 +463,15 @@ def parse_server(server):
                      'names': set()}
 
     for directive in server:
-        if directive[0] == 'listen':
+        if _is_directive(directive, 'listen'):
             addr = obj.Addr.fromstring(directive[1])
             parsed_server['addrs'].add(addr)
             if not parsed_server['ssl'] and addr.ssl:
                 parsed_server['ssl'] = True
-        elif directive[0] == 'server_name':
+        elif _is_directive(directive, 'server_name'):
             parsed_server['names'].update(
                 _get_servernames(directive[1]))
-        elif directive[0] == 'ssl' and directive[1] == 'on':
+        elif _is_directive(directive, 'ssl') and directive[1] == 'on':
             parsed_server['ssl'] = True
 
     return parsed_server
@@ -533,3 +533,7 @@ def _add_directive(block, directive, replace):
                     directive, block[location]))
         else:
             block.append(directive)
+
+def _is_directive(block, directive):
+    """Returns true if this block is of the directive type given."""
+    return len(block) > 0 and block[0] == directive
