+++ letsencrypt-nginx/letsencrypt_nginx/tests/testdata/etc_nginx/minimalistic_comments.new.conf
@@ -1,7 +1,7 @@
 # Use bar.conf when it's a full moon!
 include foo.conf;
-# Kilroy was here
 check_status;
+
 server {
     #
     # Don't forget to open up your firewall!
